+++ src/repositories/index.ts
@@ -38,3 +38,4 @@ export * from "./LocalNdjsonEventContractShadowLedgerRepository";
 export * from "./EventContractCollectionRunnerSqliteMigrationV1";
 export * from "./EventContractCollectionRunnerSqliteStore";
 export * from "./EventContractCollectionRunnerRepository";
+export * from "./EventContractCollectionRunnerSqliteRecovery";